Provide ReadableStream polyfills for Firefox and other browsers
Thanks to @jimmywarting for the inspiration; the `xhrTransport` now makes use of the `web-streams-polyfill` to return a `ReadableStream` which allows us to emulate the `Response.body()` method from the `fetch` standard.

Took this oppertunity to unify the xhr based implementations and make the transport detection be based on features rather than user-agent detection.